requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
confluent-kafka==2.6.0
clickhouse-connect==0.7.19
